+++ retroauth/src/main/java/eu/unicate/retroauth/interfaces/BaseAccountManager.java
@@ -23,8 +23,9 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import eu.unicate.retroauth.AuthenticationActivity;
+import eu.unicate.retroauth.exceptions.AuthenticationCanceledException;
 
-public interface MockableAccountManager {
+public interface BaseAccountManager {
 	/**
 	 * Gets the currently active account by the account type. The active account name is determined
 	 * by the method {@link #getActiveAccountName(String, boolean)}
@@ -80,6 +81,7 @@ public interface MockableAccountManager {
 	 * @param key         Key wiht which you want to request the value
 	 * @return The Value or <code>null</code> if the account or the key does not exist
 	 */
+	@SuppressWarnings("unused")
 	@Nullable
 	String getUserData(@NonNull String accountType, @NonNull String key);
 
@@ -126,9 +128,8 @@ public interface MockableAccountManager {
 	 * @param accountType the accountType to create if there's no account
 	 * @param tokenType   token type you need the token to be
 	 * @return the token
-	 * @throws Exception thrown when the account creation is canceled
+	 * @throws AuthenticationCanceledException thrown when the account creation is canceled
 	 */
 	@Nullable
-	String getAuthToken(@Nullable Account account, @NonNull String accountType, @NonNull String tokenType) throws Exception;
-
+	String getAuthToken(@Nullable Account account, @NonNull String accountType, @NonNull String tokenType) throws AuthenticationCanceledException;
 }